+++ src/main/java/nl/harmtacoma/autorunservices/calculator/common/Result.java
@@ -5,13 +5,13 @@ package nl.harmtacoma.autorunservices.calculator.common;
  */
 public class Result {
 
-    private int value1;
-    private int value2;
+    private float value1;
+    private float value2;
     private float result;
     private String mode;
     private String stringresult;
 
-    public Result(int value1, int value2, float result, String mode, String stringresult) {
+    public Result(float value1, float value2, float result, String mode, String stringresult) {
         setMode(mode);
         setStringresult(stringresult);
         setValue1(value1);
@@ -19,19 +19,19 @@ public class Result {
         setResult(result);
     }
 
-    public int getValue1() {
+    public float getValue1() {
         return value1;
     }
 
-    public void setValue1(int value1) {
+    public void setValue1(float value1) {
         this.value1 = value1;
     }
 
-    public int getValue2() {
+    public float getValue2() {
         return value2;
     }
 
-    public void setValue2(int value2) {
+    public void setValue2(float value2) {
         this.value2 = value2;
     }
